# Reduce allocation on StreamingHubClient broadcast events