… src/MagicOnion.Client/StreamingHubClientBase.cs